I finally added sliders and make my code more readable
