+++ Colors.py
@@ -1,11 +1,21 @@
 LIGHT_GRAY = (211, 211, 211)
+GRAY = (127, 127, 127)
 RED = (255, 0, 0)
-GREEN = (0, 255, 0)
-BLUE = (0, 0, 255)
+ORANGE = (255, 140, 0)
+GREEN_1 = (0, 255, 0)
+GREEN_2 = (127, 255, 0)
+GREEN_3 = (69, 139, 0)
+BLUE_0 = (0, 0, 255)
+BLUE_1 = (122, 197, 205)
 WHITE = (255, 255, 255)
 BLACK = (0, 0, 0)
 YELLOW = (226, 226, 29)
 PINK = (255, 90, 172)
 PURPLE = (165, 0, 255)
+BROWN = (255, 239, 219)
+AQUA = (0, 255, 255)
+AZURE = (240, 255, 255)
 
-COLORS = [RED, GREEN, BLUE, WHITE, BLACK, YELLOW, PINK, LIGHT_GRAY, PURPLE]
+
+COLORS = [BLACK, GRAY, LIGHT_GRAY, WHITE, RED, ORANGE,
+          YELLOW, GREEN_3, GREEN_1, GREEN_2,  AZURE, BLUE_0, BLUE_1, AQUA, BROWN, PURPLE, PINK]
